+++ hoomd/hpmc/compute.py
@@ -284,7 +284,7 @@ class SDF(Compute):
     where :math:`d` is the dimensionality of the system, :math:`\rho` is the
     number density, and :math:`\beta = \frac{1}{kT}`. This measurement of the
     pressure is inherently noisy due to the nature of the sampling. Average
-    `betaP` over many timesteps to obtain accurate results.
+    `P` over many timesteps to obtain accurate results.
 
     Assuming particle diameters are ~1, these parameter values typically
     achieve good results:
@@ -456,5 +456,28 @@ class SDF(Compute):
             expansion_contribution = -rho * p0_expansion / (2 * box.dimensions)
 
             return rho + compression_contribution + expansion_contribution
+        else:
+            return None
+
+    @log(requires_run=True)
+    def P(self):  # noqa: N802 - allow function name
+        """float: Pressure in NVT simulations \
+        :math:`\\left[ \\mathrm{energy} \\ \\mathrm{length}^{-d} \\right]`.
+
+        .. math::
+            P = \\frac{1}{\\beta} \\beta P
+
+        where :math:`\\beta P` is given by `betaP`.
+
+        Attention:
+            In MPI parallel execution, `P` is available on rank 0 only.
+            `P` is `None` on ranks >= 1.
+        """
+        integrator = self._simulation.operations.integrator
+
+        result = self.betaP
+
+        if result is not None:
+            return result * integrator.kT(self._simulation.timestep)
         else:
             return None