+++ hoomd/hpmc/integrate.py
@@ -28,11 +28,11 @@ not modified.
 
 .. rubric:: Temperature
 
-HPMC assumes that :math:`\beta = \frac{1}{kT} = 1`. This is not relevant to
-systems of purely hard particles where :math:`\Delta U` is either 0 or
-:math:`\infty`. To adjust the effective temperature in systems with finite
-interactions (see *Energy evaluation* below), scale the magnitude of the
-energetic interactions accordingly.
+`HPMCIntegrator.kT` (and the related :math:`\beta = \frac{1}{kT}`) are used throughout
+all HPMC operations. Set `HPMCIntegrator.kT` to control the temperature in systems with
+finite interactions (see *Energy evaluation* below). Use the default :math:`kT = 1`
+for systems of purely hard particles where :math:`\Delta U` is either 0 or
+:math:`\infty`.
 
 .. rubric:: Local trial moves
 
@@ -364,6 +364,9 @@ class HPMCIntegrator(Integrator):
         nselect (int): Number of trial moves to perform per particle per
             timestep.
 
+        kT (hoomd.variant.Variant): Temperature set point
+            :math:`[\\mathrm{energy}]`.
+
     .. rubric:: Attributes
     """
     _ext_module = _hpmc
@@ -372,14 +375,17 @@ class HPMCIntegrator(Integrator):
     _cpp_cls = None
 
     def __init__(self, default_d, default_a, translation_move_probability,
-                 nselect):
+                 nselect, kT):
         super().__init__()
 
         # Set base parameter dict for hpmc integrators
         param_dict = ParameterDict(
             translation_move_probability=float(translation_move_probability),
-            nselect=int(nselect))
+            nselect=int(nselect),
+            kT=hoomd.variant.Variant)
         self._param_dict.update(param_dict)
+        self.kT = kT
+
         self._external_potential = None
 
         # Set standard typeparameters for hpmc integrators
@@ -670,6 +676,8 @@ class Sphere(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of spheres.
     The shape :math:`S` includes all points inside and on the surface of a
@@ -721,11 +729,12 @@ class Sphere(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter('shape',
                                         type_kind='particle_types',
@@ -762,6 +771,8 @@ class ConvexPolygon(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of convex polygons. The shape :math:`S`
     of a convex polygon includes the points inside and on the surface of the
@@ -821,11 +832,12 @@ class ConvexPolygon(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter('shape',
                                         type_kind='particle_types',
@@ -862,6 +874,8 @@ class ConvexSpheropolygon(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of convex spheropolygons. The shape
     :math:`S` of a convex spheropolygon includes the points inside and on the
@@ -933,11 +947,12 @@ class ConvexSpheropolygon(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter('shape',
                                         type_kind='particle_types',
@@ -973,6 +988,8 @@ class SimplePolygon(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of simple polygons. The shape :math:`S` of
     a simple polygon includes the points inside and on the surface of the simple
@@ -1033,11 +1050,12 @@ class SimplePolygon(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter('shape',
                                         type_kind='particle_types',
@@ -1073,6 +1091,8 @@ class Polyhedron(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of general polyhedra. The shape :math:`S`
     contains the points inside the polyhedron defined by vertices and faces (see
@@ -1168,11 +1188,12 @@ class Polyhedron(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter('shape',
                                         type_kind='particle_types',
@@ -1217,6 +1238,8 @@ class ConvexPolyhedron(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of convex polyhedra. The shape :math:`S`
     of a convex polyhedron includes the points inside and on the surface of the
@@ -1271,11 +1294,12 @@ class ConvexPolyhedron(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter('shape',
                                         type_kind='particle_types',
@@ -1311,6 +1335,8 @@ class FacetedEllipsoid(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of faceted ellipsoids. The shape :math:`S`
     of a faceted ellipsoid is the intersection of an ellipsoid with a convex
@@ -1398,11 +1424,12 @@ class FacetedEllipsoid(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter('shape',
                                         type_kind='particle_types',
@@ -1435,6 +1462,8 @@ class Sphinx(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of sphere unions and differences,
     depending on the sign of the diameter. The shape :math:`S` is:
@@ -1480,11 +1509,12 @@ class Sphinx(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter('shape',
                                         type_kind='particle_types',
@@ -1508,6 +1538,8 @@ class ConvexSpheropolyhedron(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of convex spheropolyhedra. The shape
     :math:`S` of a convex spheropolyhedron includes the points inside and on the
@@ -1578,11 +1610,12 @@ class ConvexSpheropolyhedron(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter('shape',
                                         type_kind='particle_types',
@@ -1618,6 +1651,8 @@ class Ellipsoid(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of ellipsoids. The shape :math:`S`
     includes all points inside and on the surface of an ellipsoid:
@@ -1666,11 +1701,12 @@ class Ellipsoid(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter('shape',
                                         type_kind='particle_types',
@@ -1706,6 +1742,8 @@ class SphereUnion(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of unions of spheres. The union shape
     :math:`S` is the set union of the given spheres:
@@ -1776,11 +1814,12 @@ class SphereUnion(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter(
             'shape',
@@ -1835,6 +1874,8 @@ class ConvexSpheropolyhedronUnion(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of unions of convex sphereopolyhedra. The
     union shape :math:`S` is the set union of the given convex spheropolyhedra:
@@ -1915,11 +1956,12 @@ class ConvexSpheropolyhedronUnion(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter(
             'shape',
@@ -1962,6 +2004,8 @@ class FacetedEllipsoidUnion(HPMCIntegrator):
             translation moves.
         nselect (int): Number of trial moves to perform per particle per
             timestep.
+        kT (hoomd.variant.variant_like): Temperature set point
+            :math:`[\\mathrm{energy}]`.
 
     Perform hard particle Monte Carlo of unions of faceted ellipsoids. The union
     shape :math:`S` is the set union of the given faceted ellipsoids:
@@ -2065,11 +2109,12 @@ class FacetedEllipsoidUnion(HPMCIntegrator):
                  default_d=0.1,
                  default_a=0.1,
                  translation_move_probability=0.5,
-                 nselect=4):
+                 nselect=4,
+                 kT=1.0):
 
         # initialize base class
         super().__init__(default_d, default_a, translation_move_probability,
-                         nselect)
+                         nselect, kT)
 
         typeparam_shape = TypeParameter(
             'shape',